[Tests] Fix stalling Console tests due to fixture mro order change
The Console UI tests started stalling after the realese of pytest 7.4.3
which was a result of the reversing the mro order of fixtures on
classes, to give base classes priority.

The change resulted in `base_fixture` being sorted before `client` and
`daemon` fixtures and the tests stalled, likely due to the use of yield
in base_fixture.

    # pytest 7.4.2
    [Mark(name='usefixtures', args=('client', 'daemon'), kwargs={})],
    [Mark(name='usefixtures', args=('base_fixture',), kwargs={})],
    # pytest 7.4.3
    [Mark(name='usefixtures', args=('base_fixture',), kwargs={})],
    [Mark(name='usefixtures', args=('client', 'daemon'), kwargs={})],

The fix is to move base_fixture along with the other fixtures and place it
last.

Refs: https://github.com/pytest-dev/pytest/pull/11545