[Core] Refactor new features support check
In 1989d0de a check for a daemon version was added.
As this depends on knowing the version in which the feature will be
merged, this approach is not suitable for this.
Therefore, a features enum will help us list all features, without
knowing in which version each was added.